backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
pydantic-settings==2.10.1
alembic==1.14.1
psycopg[binary]==3.2.9
redis==6.4.0
celery==5.4.0
qibo==0.2.17
docker==7.1.0
pytest==8.4.1
httpx==0.28.1
requests==2.32.3
